add functional arguments, error checking

+++ strider.sh
@@ -2,12 +2,14 @@
 
 set -e
 
+
 function update() {
     echo ">> Updating"
     SCRIPT="$(dirname $0)/$(basename $0)"
     curl --silent https://raw.githubusercontent.com/panubo/strider-deploy/master/strider.sh > ${SCRIPT}.tmp
     exec bash -c "mv ${SCRIPT}.tmp ${SCRIPT} && chmod +x ${SCRIPT}"
 }
+
 
 function environment() {
     echo ">> Set Environment"
@@ -24,41 +26,68 @@ function environment() {
 
 function prepare() {
     echo ">> Preparing Checkout"
-    # TODO check for $GIT_NAME
+    if [ -z "$GIT_NAME" ]; then
+        echo "ERROR: GIT_NAME not set"
+        exit 128
+    fi
     CHECKOUT_DIR=$(basename $(pwd))
     cd ../..  # into .strider
     mkdir -p git # prepare
     rsync -a --delete data/$CHECKOUT_DIR git/$GIT_NAME # copy our clone
 }
 
+
 function test() {
     echo ">> Test Phase"
     /bin/true
 }
 
+
 function deploy() {
     echo ">> Deploying"
+
+    if [ -z "$GIT_NAME" ]; then
+        echo "ERROR: GIT_NAME not set"
+        exit 128
+    fi
+
+    if [ -z "$GIT_BRANCH" ]; then
+        echo "ERROR: GIT_BRANCH not set"
+        exit 128
+    fi
+
     UNIT="${GIT_NAME}-${GIT_BRANCH}"
     # Export Git Rev
     export GIT_HASH=$(git rev-parse HEAD)
-    echo "Git Hash: $GIT_HASH"
     # Expose ETCD vars
     for env in `tr '\0' '\n' < /proc/1/environ | grep ETCD`; do export $env; done
     # Activate Venv
     cd /data && . venv/bin/activate
     # Run Deploy
-    deploy.py --name $UNIT --method atomic --instances 2 --chunking 1 --tag ${GIT_HASH:0:7} --delay 0 --atomic-handler $(which atomic.py)
+    deploy.py --name $UNIT --method atomic --instances ${DEPLOY_INSTANCES-2} --chunking ${DEPLOY_CHUNKING-1} --tag ${GIT_HASH:0:7} --delay 0 --atomic-handler $(which atomic.py)
 }
+
 
 function cleanup() {
     echo ">> Cleanup Checkout"
-    DIR=$(pwd)
-    echo "Removing ${DIR}"
-    cd .. && rm -rf "${DIR}"
+    DIR=${1-$(pwd)}
+    cd ..
+    if [ -d "${DIR}/.git" ]; then
+        rm -rf "${DIR}"
+        echo "Removed ${DIR}"
+    else
+        echo "ERROR: Refusing to cleanup ${DIR}. Directory not found or non Git directory."
+        exit 128
+    fi
 }
+
 
 function help() {
     echo "Specify which phase to run <strider.sh> <update|environment|prepare|test|deploy|cleanup>"
 }
 
-${1-help}
+
+# Show default help. Pass arguments to the function
+CMD=${1-help}
+shift || true
+$CMD $*
